1017 원상 php 수정 : detail, update 유효성 검사

+++ src/todo_list_detail.php
@@ -139,8 +139,10 @@ try {
                         
                         <div class="btn-insert">
                             <a href="/board.php?<?php echo "page_checklist_today=".$page_checklist."&page_todo=".$page_todo ?>"><button type="button" class="btn">뒤로가기</button></a>
-                            <a href="/todo_list_delete.php?<?php echo "id=".$id."&page_checklist_today=".$page_checklist."&page_todo=".$page_todo ?>"><button type="button" class="btn">삭제하기</button></a>                            
-                            <a href="/todo_list_update.php?<?php echo "id=".$id."&page_checklist_today=".$page_checklist."&page_todo=".$page_todo ?>"><button type="button" class="btn">수정하기</button></a>
+                            <?php if($_SESSION["manager"] === true) {?>
+                                <a href="/todo_list_delete.php?<?php echo "id=".$id."&page_checklist_today=".$page_checklist."&page_todo=".$page_todo ?>"><button type="button" class="btn">삭제하기</button></a>                            
+                                <a href="/todo_list_update.php?<?php echo "id=".$id."&page_checklist_today=".$page_checklist."&page_todo=".$page_todo ?>"><button type="button" class="btn">수정하기</button></a>
+                            <?php } ?>
                             <button type="submit" class="btn">저장하기</button>
                             <input type="hidden" name="posttype" value="insert">
                         </div>
